move agentless helpers into colocated modules
