+++ x-pack/platform/plugins/shared/fleet/public/components/agentless_enrollment_flyout/index.test.tsx
@@ -13,7 +13,7 @@ import { createIntegrationsTestRendererMock } from '../../mock';
 
 import { AGENTS_PREFIX } from '../../constants';
 
-import { AgentlessEnrollmentFlyout, resolveIntegrationTitle } from '.';
+import { AgentlessEnrollmentFlyout } from '.';
 
 jest.mock('../../hooks', () => ({
   ...jest.requireActual('../../hooks'),
@@ -126,72 +126,5 @@ describe.skip('AgentlessEnrollmentFlyout', () => {
       expect(getByText('Step 2 is complete')).toBeInTheDocument();
       expect(getByText('Incoming data received from agentless integration')).toBeInTheDocument();
     });
-  });
-});
-
-describe('resolveIntegrationTitle', () => {
-  const policyTemplates = [
-    {
-      name: 'aws',
-      inputs: [
-        { type: 's3', title: 'AWS S3' },
-        { type: 'cloudwatch', title: 'AWS CloudWatch' },
-      ],
-    },
-  ];
-
-  it('falls back to the policy name when the package title is unavailable', () => {
-    expect(
-      resolveIntegrationTitle({
-        packageTitle: undefined,
-        policyTemplates,
-        selectedInput: { policyTemplate: 'aws', type: 's3' },
-        fallbackName: 'My agentless policy',
-      })
-    ).toBe('My agentless policy');
-  });
-
-  it('uses the package title when no input is selected', () => {
-    expect(
-      resolveIntegrationTitle({
-        packageTitle: 'Amazon Web Services',
-        policyTemplates,
-        selectedInput: undefined,
-        fallbackName: 'My agentless policy',
-      })
-    ).toBe('Amazon Web Services');
-  });
-
-  it('uses the selected input title when it resolves', () => {
-    expect(
-      resolveIntegrationTitle({
-        packageTitle: 'Amazon Web Services',
-        policyTemplates,
-        selectedInput: { policyTemplate: 'aws', type: 's3' },
-        fallbackName: 'My agentless policy',
-      })
-    ).toBe('AWS S3');
-  });
-
-  it('falls back to the package title when the policy template is not found', () => {
-    expect(
-      resolveIntegrationTitle({
-        packageTitle: 'Amazon Web Services',
-        policyTemplates,
-        selectedInput: { policyTemplate: 'gcp', type: 's3' },
-        fallbackName: 'My agentless policy',
-      })
-    ).toBe('Amazon Web Services');
-  });
-
-  it('falls back to the package title when the input type is not found in the template', () => {
-    expect(
-      resolveIntegrationTitle({
-        packageTitle: 'Amazon Web Services',
-        policyTemplates,
-        selectedInput: { policyTemplate: 'aws', type: 'kinesis' },
-        fallbackName: 'My agentless policy',
-      })
-    ).toBe('Amazon Web Services');
   });
 });
